requirements.txt
boto3==1.34.37
docker==7.0.0
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
pillow==10.2.0
plotly==5.18.0
pyarrow==15.0.0
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.31.1
streamlit-authenticator==0.3.1
toml==0.10.2
XlsxWriter==3.1.9
PyYAML==6.0.1
